# introduce an UncurriedPuzzle class