--- a/chia/wallet/driver_protocol.py
+++ b/chia/wallet/driver_protocol.py
@@ -12,7 +12,7 @@ class DriverProtocol(Protocol):
     def match(self, puzzle: UncurriedPuzzle) -> Optional[PuzzleInfo]:
         ...
 
-    def get_inner_puzzle(self, constructor: PuzzleInfo, puzzle_reveal: Program) -> Optional[Program]:
+    def get_inner_puzzle(self, constructor: PuzzleInfo, puzzle_reveal: UncurriedPuzzle) -> Optional[Program]:
         ...
 
     def get_inner_solution(self, constructor: PuzzleInfo, solution: Program) -> Optional[Program]: