--- a/chia/wallet/nft_wallet/singleton_outer_puzzle.py
+++ b/chia/wallet/nft_wallet/singleton_outer_puzzle.py
@@ -14,17 +14,18 @@ from chia.wallet.puzzles.singleton_top_layer_v1_1 import (
     puzzle_for_singleton,
     solution_for_singleton,
 )
+from chia.wallet.uncurried_puzzle import UncurriedPuzzle, uncurry_puzzle
 
 
 @dataclass(frozen=True)
 class SingletonOuterPuzzle:
-    _match: Callable[[Program], Optional[PuzzleInfo]]
+    _match: Callable[[UncurriedPuzzle], Optional[PuzzleInfo]]
     _construct: Callable[[PuzzleInfo, Program], Program]
     _solve: Callable[[PuzzleInfo, Solver, Program, Program], Program]
     _get_inner_puzzle: Callable[[PuzzleInfo, Program], Optional[Program]]
     _get_inner_solution: Callable[[PuzzleInfo, Program], Optional[Program]]
 
-    def match(self, puzzle: Program) -> Optional[PuzzleInfo]:
+    def match(self, puzzle: UncurriedPuzzle) -> Optional[PuzzleInfo]:
         matched, curried_args = match_singleton_puzzle(puzzle)
         if matched:
             singleton_struct, inner_puzzle = curried_args
@@ -33,7 +34,7 @@ class SingletonOuterPuzzle:
                 "launcher_id": "0x" + singleton_struct.as_python()[1].hex(),
                 "launcher_ph": "0x" + singleton_struct.as_python()[2].hex(),
             }
-            next_constructor = self._match(inner_puzzle)
+            next_constructor = self._match(uncurry_puzzle(inner_puzzle))
             if next_constructor is not None:
                 constructor_dict["also"] = next_constructor.info
             return PuzzleInfo(constructor_dict)
@@ -51,7 +52,7 @@ class SingletonOuterPuzzle:
         return puzzle_for_singleton(constructor["launcher_id"], inner_puzzle, launcher_hash)
 
     def get_inner_puzzle(self, constructor: PuzzleInfo, puzzle_reveal: Program) -> Optional[Program]:
-        matched, curried_args = match_singleton_puzzle(puzzle_reveal)
+        matched, curried_args = match_singleton_puzzle(uncurry_puzzle(puzzle_reveal))
         if matched:
             _, inner_puzzle = curried_args
             also = constructor.also()
@@ -80,7 +81,7 @@ class SingletonOuterPuzzle:
         also = constructor.also()
         if also is not None:
             inner_solution = self._solve(also, solver, inner_puzzle, inner_solution)
-        matched, curried_args = match_singleton_puzzle(parent_spend.puzzle_reveal.to_program())
+        matched, curried_args = match_singleton_puzzle(uncurry_puzzle(parent_spend.puzzle_reveal.to_program()))
         assert matched
         _, parent_inner_puzzle = curried_args
         return solution_for_singleton(