--- a/chia/wallet/nft_wallet/transfer_program_puzzle.py
+++ b/chia/wallet/nft_wallet/transfer_program_puzzle.py
@@ -7,14 +7,14 @@ from chia.util.ints import uint16
 from chia.wallet.puzzle_drivers import PuzzleInfo, Solver
 from chia.wallet.puzzles.load_clvm import load_clvm
 from chia.wallet.puzzles.singleton_top_layer_v1_1 import SINGLETON_LAUNCHER_HASH, SINGLETON_MOD_HASH
+from chia.wallet.uncurried_puzzle import UncurriedPuzzle
 
 TRANSFER_PROGRAM_MOD = load_clvm("nft_ownership_transfer_program_one_way_claim_with_royalties.clvm")
 
 
-def match_transfer_program_puzzle(puzzle: Program) -> Tuple[bool, List[Program]]:
-    mod, args = puzzle.uncurry()
-    if mod == TRANSFER_PROGRAM_MOD:
-        return True, list(args.as_iter())
+def match_transfer_program_puzzle(puzzle: UncurriedPuzzle) -> Tuple[bool, List[Program]]:
+    if puzzle.mod == TRANSFER_PROGRAM_MOD:
+        return True, list(puzzle.args.as_iter())
     return False, []
 
 
@@ -39,13 +39,13 @@ def solution_for_transfer_program(
 
 @dataclass(frozen=True)
 class TransferProgramPuzzle:
-    _match: Callable[[Program], Optional[PuzzleInfo]]
+    _match: Callable[[UncurriedPuzzle], Optional[PuzzleInfo]]
     _construct: Callable[[PuzzleInfo, Program], Program]
     _solve: Callable[[PuzzleInfo, Solver, Program, Program], Program]
     _get_inner_puzzle: Callable[[PuzzleInfo, Program], Optional[Program]]
     _get_inner_solution: Callable[[PuzzleInfo, Program], Optional[Program]]
 
-    def match(self, puzzle: Program) -> Optional[PuzzleInfo]:
+    def match(self, puzzle: UncurriedPuzzle) -> Optional[PuzzleInfo]:
         matched, curried_args = match_transfer_program_puzzle(puzzle)
         if matched:
             singleton_struct, royalty_puzzle_hash, percentage = curried_args