--- a/chia/wallet/nft_wallet/transfer_program_puzzle.py
+++ b/chia/wallet/nft_wallet/transfer_program_puzzle.py
@@ -42,7 +42,7 @@ class TransferProgramPuzzle:
     _match: Callable[[UncurriedPuzzle], Optional[PuzzleInfo]]
     _construct: Callable[[PuzzleInfo, Program], Program]
     _solve: Callable[[PuzzleInfo, Solver, Program, Program], Program]
-    _get_inner_puzzle: Callable[[PuzzleInfo, Program], Optional[Program]]
+    _get_inner_puzzle: Callable[[PuzzleInfo, UncurriedPuzzle], Optional[Program]]
     _get_inner_solution: Callable[[PuzzleInfo, Program], Optional[Program]]
 
     def match(self, puzzle: UncurriedPuzzle) -> Optional[PuzzleInfo]:
@@ -67,7 +67,7 @@ class TransferProgramPuzzle:
             constructor["launcher_id"], constructor["royalty_address"], constructor["royalty_percentage"]
         )
 
-    def get_inner_puzzle(self, constructor: PuzzleInfo, puzzle_reveal: Program) -> Optional[Program]:
+    def get_inner_puzzle(self, constructor: PuzzleInfo, puzzle_reveal: UncurriedPuzzle) -> Optional[Program]:
         return None
 
     def get_inner_solution(self, constructor: PuzzleInfo, solution: Program) -> Optional[Program]: