--- a/chia/wallet/outer_puzzles.py
+++ b/chia/wallet/outer_puzzles.py
@@ -10,12 +10,13 @@ from chia.wallet.nft_wallet.ownership_outer_puzzle import OwnershipOuterPuzzle
 from chia.wallet.nft_wallet.singleton_outer_puzzle import SingletonOuterPuzzle
 from chia.wallet.nft_wallet.transfer_program_puzzle import TransferProgramPuzzle
 from chia.wallet.puzzle_drivers import PuzzleInfo, Solver
+from chia.wallet.uncurried_puzzle import UncurriedPuzzle
 
 """
 This file provides a central location for acquiring drivers for outer puzzles like CATs, NFTs, etc.
 
 A driver for a puzzle must include the following functions:
-  - match(self, puzzle: Program) -> Optional[PuzzleInfo]
+  - match(self, puzzle: UncurriedPuzzle) -> Optional[PuzzleInfo]
     - Given a puzzle reveal, return a PuzzleInfo object that can be used to reconstruct it later
   - get_inner_puzzle(self, constructor: PuzzleInfo, puzzle_reveal: Program) -> Optional[Program]:
     - Given a PuzzleInfo object and a puzzle reveal, pull out this outer puzzle's inner puzzle
@@ -39,7 +40,7 @@ class AssetType(Enum):
     ROYALTY_TRANSFER_PROGRAM = "royalty transfer program"
 
 
-def match_puzzle(puzzle: Program) -> Optional[PuzzleInfo]:
+def match_puzzle(puzzle: UncurriedPuzzle) -> Optional[PuzzleInfo]:
     for driver in driver_lookup.values():
         potential_info: Optional[PuzzleInfo] = driver.match(puzzle)
         if potential_info is not None: