… chia/wallet/puzzles/singleton_top_layer_v1_1.py
… chia/wallet/trading/offer.py
… chia/wallet/wallet_state_manager.py
… tests/wallet/cat_wallet/test_cat_outer_puzzle.py
… tests/wallet/nft_wallet/test_nft_1_offers.py
… tests/wallet/nft_wallet/test_nft_offers.py
… tests/wallet/nft_wallet/test_nft_puzzles.py
… tests/wallet/nft_wallet/test_ownership_outer_puzzle.py
… tools/run_block.py